initializeing...

+++ DAL/Drone.cs
@@ -17,7 +17,7 @@ namespace IDAL
             public string Model { get; set; }
             public @enum.WeightCategories Weight { get; set; }
             public @enum.DroneStatus Status { get; set; }
-            public double Battery { get; set; }
+            public int Battery { get; set; }
             public @enum.Priorities Priority { get; set; }
         }
     }
